Make the code less bad
