+++ dbms/src/DataStreams/PushingToViewsBlockOutputStream.h
@@ -5,7 +5,6 @@
 #include <DataStreams/OneBlockInputStream.h>
 #include <DataStreams/MaterializingBlockInputStream.h>
 #include <Storages/StorageMaterializedView.h>
-#include <Storages/StorageLiveView.h>
 
 namespace DB
 {
