+++ dbms/src/Storages/LiveView/LiveViewEventsBlockInputStream.h
@@ -9,11 +9,9 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
+
 #pragma once
 
-#include <limits>
-
-#include <Common/ConcurrentBoundedQueue.h>
 #include <Poco/Condition.h>
 #include <DataTypes/DataTypesNumber.h>
 #include <DataTypes/DataTypeString.h>
@@ -21,7 +19,7 @@ limitations under the License. */
 #include <Columns/ColumnsNumber.h>
 #include <DataStreams/OneBlockInputStream.h>
 #include <DataStreams/IBlockInputStream.h>
-#include <Storages/StorageLiveView.h>
+#include <Storages/LiveView/StorageLiveView.h>
 
 
 namespace DB
@@ -66,8 +64,8 @@ public:
         if (isCancelled() || storage->is_dropped)
             return;
         IBlockInputStream::cancel(kill);
-        Poco::FastMutex::ScopedLock lock(storage->mutex);
-        storage->condition.broadcast();
+        std::lock_guard lock(storage->mutex);
+        storage->condition.notify_all();
     }
 
     Block getHeader() const override
@@ -103,7 +101,7 @@ public:
 
     NonBlockingResult tryRead()
     {
-        return tryRead_(false);
+        return tryReadImpl(false);
     }
 
     Block getEventBlock()
@@ -120,7 +118,7 @@ protected:
     Block readImpl() override
     {
         /// try reading
-        return tryRead_(true).first;
+        return tryReadImpl(true).first;
     }
 
     /** tryRead method attempts to read a block in either blocking
@@ -128,7 +126,7 @@ protected:
      *  then method return empty block with flag set to false
      *  to indicate that method would block to get the next block.
      */
-    NonBlockingResult tryRead_(bool blocking)
+    NonBlockingResult tryReadImpl(bool blocking)
     {
         if (has_limit && num_updates == static_cast<Int64>(limit))
         {
@@ -137,7 +135,7 @@ protected:
         /// If blocks were never assigned get blocks
         if (!blocks)
         {
-            Poco::FastMutex::ScopedLock lock(storage->mutex);
+            std::lock_guard lock(storage->mutex);
             if (!active)
                 return { Block(), false };
             blocks = (*blocks_ptr);
@@ -155,7 +153,7 @@ protected:
         if (it == end)
         {
             {
-                Poco::FastMutex::ScopedLock lock(storage->mutex);
+                std::unique_lock lock(storage->mutex);
                 if (!active)
                     return { Block(), false };
                 /// If we are done iterating over our blocks
@@ -183,7 +181,10 @@ protected:
                     while (true)
                     {
                         UInt64 timestamp_usec = static_cast<UInt64>(timestamp.epochMicroseconds());
-                        bool signaled = storage->condition.tryWait(storage->mutex, std::max(static_cast<UInt64>(0), heartbeat_interval_usec - (timestamp_usec - last_event_timestamp_usec)) / 1000);
+
+                        /// Or spurious wakeup.
+                        bool signaled = std::cv_status::no_timeout == storage->condition.wait_for(lock,
+                            std::chrono::microseconds(std::max(UInt64(0), heartbeat_interval_usec - (timestamp_usec - last_event_timestamp_usec))));
 
                         if (isCancelled() || storage->is_dropped)
                         {
@@ -202,7 +203,7 @@ protected:
                     }
                 }
             }
-            return tryRead_(blocking);
+            return tryReadImpl(blocking);
         }
 
         // move right to the end
