+++ dbms/src/Storages/StorageFactory.cpp
@@ -68,14 +68,6 @@ StoragePtr StorageFactory::get(
 
         name = "LiveView";
     }
-    else if (query.is_live_channel)
-    {
-
-        if (query.storage)
-            throw Exception("Specifying ENGINE is not allowed for a LiveChannel", ErrorCodes::INCORRECT_QUERY);
-
-        name = "LiveChannel";
-    }
     else
     {
         /// Check for some special types, that are not allowed to be stored in tables. Example: NULL data type.
@@ -135,12 +127,6 @@ StoragePtr StorageFactory::get(
             {
                 throw Exception(
                     "Direct creation of tables with ENGINE LiveView is not supported, use CREATE LIVE VIEW statement",
-                    ErrorCodes::INCORRECT_QUERY);
-            }
-            else if (name == "LiveChannel")
-            {
-                throw Exception(
-                    "Direct creation of tables with ENGINE LiveChannel is not supported, use CREATE LIVE CHANNEL statement",
                     ErrorCodes::INCORRECT_QUERY);
             }
         }
